move context menu callbacks into component iteself
